Added relief, maxDEM and minDEM in 'get_cov'. Added 'singletons' to 'get_grid'.  Added occupancy drop warning to 'SubSamp_for_unmarked'
Added various small extras to functions. Relief, max and min DEM values added to allow for inclusion of relief in covariates. Singletons added to allow the user to easily explore which collections/occurrences are singletons within gridcells. Warning added to alert the user to a drop in naive occupancy when using subsampling option - provides number of sites dropped, approximate % of sites reduced in occupancy to 0, and a dataframe that compares occupancy of subsampled to original sites.